split cells by offset (new algorithm) multiple materials in a single cell (new data structure) material assignments using typed references CellContainer passing/storing using shared_ptr to avoid copy from int to size_t for some positions adding to_str() to CellContainer and NodeContainer specific data structures for a given cell type (work in progress) fix bug PCOMP (missing cell group assignment) right value assignment optimisation in SystusWriter unit test (nastran) for multiline COOR2C

+++ Aster/MedWriter.cpp
@@ -92,7 +92,7 @@ vector<int> NodeGroup2Families::getFamilyOnNodes() const {
 }
 
 CellGroup2Families::CellGroup2Families(
-		const Mesh& mesh, unordered_map<CellType::Code, int, EnumClassHash> cellCountByType,
+		const Mesh& mesh, unordered_map<CellType::Code, size_t, EnumClassHash> cellCountByType,
 		const vector<shared_ptr<CellGroup>>& cellGroups) : mesh(mesh) {
 	int currentFamilyId = 0;
 	unordered_map<int, int> newFamilyByOldfamily;
@@ -320,9 +320,9 @@ void MedWriter::writeMED(const Model& model, const string& medFileName) {
 	}
 	vector<shared_ptr<CellGroup>> cellGroups = model.mesh.getCellGroups();
 	if (not cellGroups.empty()) {
-		unordered_map<CellType::Code, int, EnumClassHash> cellCountByType;
+		unordered_map<CellType::Code, size_t, EnumClassHash> cellCountByType;
 		for (const auto& typeAndCodePair : CellType::typeByCode) {
-			int cellNum = model.mesh.countCells(*typeAndCodePair.second);
+			size_t cellNum = model.mesh.countCells(*typeAndCodePair.second);
 			if (cellNum > 0) {
 				cellCountByType[typeAndCodePair.first] = cellNum;
 			}
